[extract-form-example] fix form example
